+++ lib/flutter_inappbrowser.dart
@@ -61,66 +61,63 @@ class InAppBrowser {
 
   ///Opens an [url] in a new [InAppBrowser] instance or the system browser.
   ///
-  /// - [url]: The [url] to load __(String)__. Call [encodeUriComponent()] on this if the [url] contains Unicode characters.
+  ///- [url]: The [url] to load. Call [encodeUriComponent()] on this if the [url] contains Unicode characters.
   ///
-  /// - [target]: The target in which to load the [url], an optional parameter that defaults to `_self`. __(String)__
+  ///- [headers]: The additional headers to be used in the HTTP request for this URL, specified as a map from name to value.
   ///
-  ///   - `_self`: Opens in the [InAppBrowser].
-  ///   - `_blank`: Opens in the [InAppBrowser].
-  ///   - `_system`: Opens in the system's web browser.
+  ///- [target]: The target in which to load the [url], an optional parameter that defaults to `_self`.
   ///
-  ///   - [options]: Options for the `InAppBrowser`. Optional, defaulting to: `location=yes`. _(String)_
+  ///  - `_self`: Opens in the [InAppBrowser].
+  ///  - `_blank`: Opens in the [InAppBrowser].
+  ///  - `_system`: Opens in the system's web browser.
   ///
-  ///   The [options] string must not contain any blank space, and each feature's name/value pairs must be separated by a comma. Feature names are case insensitive.
+  ///- [options]: Options for the [InAppBrowser].
   ///
-  ///   All platforms support:
+  ///  All platforms support:
+  ///  - __useShouldOverrideUrlLoading__: Set to `true` to be able to listen at the [shouldOverrideUrlLoading()] event. The default value is `false`.
+  ///  - __clearCache__: Set to `true` to have all the browser's cache cleared before the new window is opened. The default value is `false`.
+  ///  - __userAgent___: Set the custom WebView's user-agent.
+  ///  - __javaScriptEnabled__: Set to `true` to enable JavaScript. The default value is `true`.
+  ///  - __javaScriptCanOpenWindowsAutomatically__: Set to `true` to allow JavaScript open windows without user interaction. The default value is `false`.
+  ///  - __hidden__: Set to `true` to create the browser and load the page, but not show it. The `onLoadStop` event fires when loading is complete. Omit or set to `false` (default) to have the browser open and load normally.
+  ///  - __toolbarTop__: Set to `false` to hide the toolbar at the top of the WebView. The default value is `true`.
+  ///  - __toolbarTopBackgroundColor__: Set the custom background color of the toolbat at the top.
+  ///  - __hideUrlBar__: Set to `true` to hide the url bar on the toolbar at the top. The default value is `false`.
+  ///  - __mediaPlaybackRequiresUserGesture__: Set to `true` to prevent HTML5 audio or video from autoplaying. The default value is `true`.
   ///
-  ///   - __location__: Set to `yes` or `no` to turn the [InAppBrowser]'s location bar on or off.
+  ///  **Android** supports these additional options:
   ///
-  ///   **Android** supports these additional options:
+  ///  - __hideTitleBar__: Set to `true` if you want the title should be displayed. The default value is `false`.
+  ///  - __closeOnCannotGoBack__: Set to `false` to not close the InAppBrowser when the user click on the back button and the WebView cannot go back to the history. The default value is `true`.
+  ///  - __clearSessionCache__: Set to `true` to have the session cookie cache cleared before the new window is opened.
+  ///  - __builtInZoomControls__: Set to `true` if the WebView should use its built-in zoom mechanisms. The default value is `false`.
+  ///  - __supportZoom__: Set to `false` if the WebView should not support zooming using its on-screen zoom controls and gestures. The default value is `true`.
+  ///  - __databaseEnabled__: Set to `true` if you want the database storage API is enabled. The default value is `false`.
+  ///  - __domStorageEnabled__: Set to `true` if you want the DOM storage API is enabled. The default value is `false`.
+  ///  - __useWideViewPort__: Set to `true` if the WebView should enable support for the "viewport" HTML meta tag or should use a wide viewport. When the value of the setting is false, the layout width is always set to the width of the WebView control in device-independent (CSS) pixels. When the value is true and the page contains the viewport meta tag, the value of the width specified in the tag is used. If the page does not contain the tag or does not provide a width, then a wide viewport will be used. The default value is `true`.
+  ///  - __safeBrowsingEnabled__: Set to `true` if you want the Safe Browsing is enabled. Safe Browsing allows WebView to protect against malware and phishing attacks by verifying the links. The default value is `true`.
+  ///  - __progressBar__: Set to `false` to hide the progress bar at the bottom of the toolbar at the top. The default value is `true`.
   ///
-  ///   - __hidden__: set to `yes` to create the browser and load the page, but not show it. The loadstop event fires when loading is complete. Omit or set to `no` (default) to have the browser open and load normally.
-  ///   - __clearcache__: set to `yes` to have the browser's cookie cache cleared before the new window is opened
-  ///   - __clearsessioncache__: set to `yes` to have the session cookie cache cleared before the new window is opened
-  ///   - __closebuttoncaption__: set to a string to use as the close button's caption instead of a X. Note that you need to localize this value yourself.
-  ///   - __closebuttoncolor__: set to a valid hex color string, for example: `#00ff00`, and it will change the
-  ///   close button color from default, regardless of being a text or default X. Only has effect if user has location set to `yes`.
-  ///   - __footer__: set to `yes` to show a close button in the footer similar to the iOS __Done__ button.
-  ///   The close button will appear the same as for the header hence use __closebuttoncaption__ and __closebuttoncolor__ to set its properties.
-  ///   - __footercolor__: set to a valid hex color string, for example `#00ff00` or `#CC00ff00` (`#aarrggbb`) , and it will change the footer color from default.
-  ///   Only has effect if user has __footer__ set to `yes`.
-  ///   - __hardwareback__: set to `yes` to use the hardware back button to navigate backwards through the `InAppBrowser`'s history. If there is no previous page, the `InAppBrowser` will close.  The default value is `yes`, so you must set it to `no` if you want the back button to simply close the InAppBrowser.
-  ///   - __hidenavigationbuttons__: set to `yes` to hide the navigation buttons on the location toolbar, only has effect if user has location set to `yes`. The default value is `no`.
-  ///   - __hideurlbar__: set to `yes` to hide the url bar on the location toolbar, only has effect if user has location set to `yes`. The default value is `no`.
-  ///   - __navigationbuttoncolor__: set to a valid hex color string, for example: `#00ff00`, and it will change the color of both navigation buttons from default. Only has effect if user has location set to `yes` and not hidenavigationbuttons set to `yes`.
-  ///   - __toolbarcolor__: set to a valid hex color string, for example: `#00ff00`, and it will change the color the toolbar from default. Only has effect if user has location set to `yes`.
-  ///   - __zoom__: set to `yes` to show Android browser's zoom controls, set to `no` to hide them.  Default value is `yes`.
-  ///   - __mediaPlaybackRequiresUserAction__: Set to `yes` to prevent HTML5 audio or video from autoplaying (defaults to `no`).
-  ///   - __shouldPauseOnSuspend__: Set to `yes` to make InAppBrowser WebView to pause/resume with the app to stop background audio (this may be required to avoid Google Play issues like described in [CB-11013](https://issues.apache.org/jira/browse/CB-11013)).
-  ///   - __useWideViewPort__: Sets whether the WebView should enable support for the "viewport" HTML meta tag or should use a wide viewport. When the value of the setting is `no`, the layout width is always set to the width of the WebView control in device-independent (CSS) pixels. When the value is `yes` and the page contains the viewport meta tag, the value of the width specified in the tag is used. If the page does not contain the tag or does not provide a width, then a wide viewport will be used. (defaults to `yes`).
+  ///  **iOS** supports these additional options:
   ///
-  ///   **iOS** supports these additional options:
-  ///
-  ///   - __hidden__: set to `yes` to create the browser and load the page, but not show it. The loadstop event fires when loading is complete. Omit or set to `no` (default) to have the browser open and load normally.
-  ///   - __clearcache__: set to `yes` to have the browser's cookie cache cleared before the new window is opened
-  ///   - __clearsessioncache__: set to `yes` to have the session cookie cache cleared before the new window is opened
-  ///   - __closebuttoncolor__: set as a valid hex color string, for example: `#00ff00`, to change from the default __Done__ button's color. Only applicable if toolbar is not disabled.
-  ///   - __closebuttoncaption__: set to a string to use as the __Done__ button's caption. Note that you need to localize this value yourself.
-  ///   - __disallowoverscroll__: Set to `yes` or `no` (default is `no`). Turns on/off the UIWebViewBounce property.
-  ///   - __hidenavigationbuttons__:  set to `yes` or `no` to turn the toolbar navigation buttons on or off (defaults to `no`). Only applicable if toolbar is not disabled.
-  ///   - __navigationbuttoncolor__:  set as a valid hex color string, for example: `#00ff00`, to change from the default color. Only applicable if navigation buttons are visible.
-  ///   - __toolbar__:  set to `yes` or `no` to turn the toolbar on or off for the InAppBrowser (defaults to `yes`)
-  ///   - __toolbarcolor__: set as a valid hex color string, for example: `#00ff00`, to change from the default color of the toolbar. Only applicable if toolbar is not disabled.
-  ///   - __toolbartranslucent__:  set to `yes` or `no` to make the toolbar translucent(semi-transparent)  (defaults to `yes`). Only applicable if toolbar is not disabled.
-  ///   - __enableViewportScale__:  Set to `yes` or `no` to prevent viewport scaling through a meta tag (defaults to `no`).
-  ///   - __mediaPlaybackRequiresUserAction__: Set to `yes` to prevent HTML5 audio or video from autoplaying (defaults to `no`).
-  ///   - __allowInlineMediaPlayback__: Set to `yes` or `no` to allow in-line HTML5 media playback, displaying within the browser window rather than a device-specific playback interface. The HTML's `video` element must also include the `webkit-playsinline` attribute (defaults to `no`)
-  ///   - __keyboardDisplayRequiresUserAction__: Set to `yes` or `no` to open the keyboard when form elements receive focus via JavaScript's `focus()` call (defaults to `yes`).
-  ///   - __suppressesIncrementalRendering__: Set to `yes` or `no` to wait until all new view content is received before being rendered (defaults to `no`).
-  ///   - __presentationstyle__:  Set to `pagesheet`, `formsheet` or `fullscreen` to set the [presentation style](http://developer.apple.com/library/ios/documentation/UIKit/Reference/UIViewController_Class/Reference/Reference.html#//apple_ref/occ/instp/UIViewController/modalPresentationStyle) (defaults to `fullscreen`).
-  ///   - __transitionstyle__: Set to `fliphorizontal`, `crossdissolve` or `coververtical` to set the [transition style](http://developer.apple.com/library/ios/#documentation/UIKit/Reference/UIViewController_Class/Reference/Reference.html#//apple_ref/occ/instp/UIViewController/modalTransitionStyle) (defaults to `coververtical`).
-  ///   - __toolbarposition__: Set to `top` or `bottom` (default is `bottom`). Causes the toolbar to be at the top or bottom of the window.
-  ///   - __hidespinner__: Set to `yes` or `no` to change the visibility of the loading indicator (defaults to `no`).
+  ///  - __disallowOverScroll__: Set to `true` to disable the bouncing of the WebView when the scrolling has reached an edge of the content. The default value is `false`.
+  ///  - __toolbarBottom__: Set to `false` to hide the toolbar at the bottom of the WebView. The default value is `true`.
+  ///  - __toolbarBottomBackgroundColor__: Set the custom background color of the toolbat at the bottom.
+  ///  - __toolbarBottomTranslucent__: Set to `true` to set the toolbar at the bottom translucent. The default value is `true`.
+  ///  - __closeButtonCaption__: Set the custom text for the close button.
+  ///  - __closeButtonColor__: Set the custom color for the close button.
+  ///  - __presentationStyle__: Set the custom modal presentation style when presenting the WebView. The default value is `0 //fullscreen`. See [UIModalPresentationStyle](https://developer.apple.com/documentation/uikit/uimodalpresentationstyle) for all the available styles.
+  ///  - __transitionStyle__: Set to the custom transition style when presenting the WebView. The default value is `0 //crossDissolve`. See [UIModalTransitionStyle](https://developer.apple.com/documentation/uikit/uimodaltransitionStyle) for all the available styles.
+  ///  - __enableViewportScale__: Set to `true` to allow a viewport meta tag to either disable or restrict the range of user scaling. The default value is `false`.
+  ///  - __keyboardDisplayRequiresUserAction__: Set to `true` if you want the user must explicitly tap the elements in the WebView to display the keyboard (or other relevant input view) for that element. When set to `false`, a focus event on an element causes the input view to be displayed and associated with that element automatically. The default value is `true`.
+  ///  - __suppressesIncrementalRendering__: Set to `true` if you want the WebView suppresses content rendering until it is fully loaded into memory.. The default value is `false`.
+  ///  - __allowsAirPlayForMediaPlayback__: Set to `true` to allow AirPlay. The default value is `true`.
+  ///  - __allowsBackForwardNavigationGestures__: Set to `true` to allow the horizontal swipe gestures trigger back-forward list navigations. The default value is `true`.
+  ///  - __allowsLinkPreview__: Set to `true` to allow that pressing on a link displays a preview of the destination for the link. The default value is `true`.
+  ///  - __ignoresViewportScaleLimits__: Set to `true` if you want that the WebView should always allow scaling of the webpage, regardless of the author's intent. The ignoresViewportScaleLimits property overrides the `user-scalable` HTML property in a webpage. The default value is `false`.
+  ///  - __allowsInlineMediaPlayback__: Set to `true` to allow HTML5 media playback to appear inline within the screen layout, using browser-supplied controls rather than native controls. For this to work, add the `webkit-playsinline` attribute to any `<video>` elements. The default value is `false`.
+  ///  - __allowsPictureInPictureMediaPlayback__: Set to `true` to allow HTML5 videos play picture-in-picture. The default value is `true`.
+  ///  - __spinner__: Set to `false` to hide the spinner when the WebView is loading a page. The default value is `true`.
   Future<void> open(String url, {Map<String, String> headers = const {}, String target = "_self", Map<String, dynamic> options = const {}}) async {
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent('url', () => url);
@@ -130,6 +127,7 @@ class InAppBrowser {
     return await _channel.invokeMethod('open', args);
   }
 
+  ///Loads the given [url] with optional [headers] specified as a map from name to value.
   Future<void> loadUrl(String url, {Map<String, String> headers = const {}}) async {
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent('url', () => url);
@@ -182,7 +180,7 @@ class InAppBrowser {
     return await _channel.invokeMethod('isHidden');
   }
 
-  ///Injects JavaScript code into the [InAppBrowser] window. (Only available when the target is set to `_blank` or to `_self`)
+  ///Injects JavaScript code into the [InAppBrowser] window and returns the result of the evaluation. (Only available when the target is set to `_blank` or to `_self`)
   Future<String> injectScriptCode(String source) async {
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent('source', () => source);
@@ -230,6 +228,7 @@ class InAppBrowser {
 
   }
 
+  ///Give the host application a chance to take control when a URL is about to be loaded in the current WebView.
   void shouldOverrideUrlLoading(String url) {
 
   }