+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.0
+
+- updated/added new methods
+- updated UI of android/iOS in-app browser
+- code cleanup
+- added new options when opening the in-app browser
+
 ## 0.0.1
 
 Initial release.